Second navbar completed, fixed with scroll option
that changes to a transparent navbar when scrolled on top of the picture
and changes back to a solid color when it reaches 500px in the page.
Side menu completed with the options it will have. Still need to link
the actual pages it will direct to to it. Search bar is not working,
because I haven't worked on it yet. Next will be the blog displayed on
the page.